[DAD-782] refactor: navigation 파일 모으기
