+++ src/pages/ScrapPage.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import Navbar from '../components/molcules/Navbar';
+import Navbar from '../components/molcules/Navigation/Navbar';
 import { Outlet } from 'react-router-dom';
 
 function ScrapPage() {
